requirements.txt
fastapi[all]>=0.116.1,<0.117.0
uvicorn[standard]==0.35.0
SQLAlchemy[asyncio]>=2.0.42,<2.1.0
asyncpg==0.30.0
anyio==4.10.0
async-timeout==5.0.1
aiofiles==23.1.0
httpx==0.28.1
jinja2>=3.1.5,<4.0.0
python-dotenv==1.1.1
pydantic==2.11.7
watchfiles==1.1.0
requests==2.32.4
typing_extensions==4.14.1
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
sniffio==1.3.1
urllib3==2.5.0